Moved client from a local variable to module namespace.
Now instead of needing to created and hold onto an instance of ElasticSearch::Client throughout an application, it's stored in a singleton in the ElasticSearch namespace.

To setup, one calls ElasticSearch.setup(servers, options), and then you have access to the client at ElasticSearch.client.

There's some cleaning up to be done here. For one, it may be better to do the setup in a block, which will be something I look at. Secondly this may not be thread safe. I've not looked sufficiently through the innards. Thirdly, there's just some a little safety required (like properly handling ElasticSearch.client being accessed before ElasticSearch.setup is called.